Final merge

+++ priv/repo/migrations/20191218210507_create_accounts.exs
@@ -3,7 +3,7 @@ defmodule Bank.Repo.Migrations.CreateAccounts do
 
   def change do
     create table(:accounts) do
-      add :balance, :decimal, null: false
+      add :balance, :float, null: false
       add :code, :string, null: false
       add :digit, :integer, null: false
       add :agency_id, references(:agencies), null: false
